requirements.txt
asgiref==3.5.2
image==1.5.33
imutils==0.5.4
numpy==1.23.5
opencv-python==4.6.0.66
Pillow==9.3.0
pytz==2022.6
sqlparse==0.4.3